Style Nest page
