+++ src/index.css
@@ -33,23 +33,23 @@ a:hover, a:focus {
 }
 
 .tooltip > .tooltip-inner {
-  background-color: #333;
+  background-color: #555;
 }
 
 .tooltip.bs-tooltip-left > .tooltip-arrow::before {
-  border-left-color: #333;
+  border-left-color: #555;
 }
 
 .tooltip.bs-tooltip-right > .tooltip-arrow::before {
-  border-right-color: #333;
+  border-right-color: #555;
 }
 
 .tooltip.bs-tooltip-top > .tooltip-arrow::before {
-  border-top-color: #333;
+  border-top-color: #555;
 }
 
 .tooltip.bs-tooltip-bottom > .tooltip-arrow::before {
-  border-bottom-color: #333;
+  border-bottom-color: #555;
 }
 
 ::selection {
